Add developer.here Places API

+++ src/main/scala/model/SearchRequest.scala
@@ -2,4 +2,4 @@ package model
 
 import com.bot4s.telegram.models.Location
 
-final case class SearchRequest(place: PlaceType, location: Option[Location] = None, radius: Long = 1000)
+final case class SearchRequest(place: PlaceType, location: Option[Location] = None, radius: Double = 1000)
